+++ src/Sql/Sql/Replication/Model/AzureReplicationLinkModel.cs
@@ -113,6 +113,16 @@ namespace Microsoft.Azure.Commands.Sql.Replication.Model
         public string LicenseType { get; set; }
 
         /// <summary>
+        /// Gets or sets the Auto pause delay of the database
+        /// </summary>
+        public int? AutoPauseDelayInMinutes { get; set; }
+
+        /// <summary>
+        /// Minimal capacity that database will always have allocated, if not paused
+        /// </summary>
+        public double? MinimumCapacity { get; set; }
+
+        /// <summary>
         /// Gets or sets the current backup storage redundancy for the database
         /// </summary>
         public string CurrentBackupStorageRedundancy { get; set; }