+++ src/Compute/Compute/ChangeLog.md
@@ -22,6 +22,7 @@
 ## Upcoming Release
 * Fixed `New-AzVmss` to correctly work when using `-EdgeZone` by creating the Load Balancer in the correct edge zone.
 * Removed references to image aliases in `New-AzVM` and `New-AzVmss` to images that were removed.
+* Az.Compute is updated to use the 2023-09-01 ComputeRP REST API calls. 
 
 ## Version 7.1.0
 * Added new parameter `-ElasticSanResourceId` to `New-AzSnapshotConfig` cmdlet.